scraper/requirements.txt
playwright    ==1.49.1
pymysql       ==1.1.1
sqlalchemy    ==2.0.36
cryptography  ==44.0.0
python-dotenv ==1.0.1
httpx         ==0.28.1